refactor: rename and cover lumberjack-log-driver.ts
